remove int32 convrsion

+++ internal/otelcollector/ports/ports.go
@@ -1,10 +1,10 @@
 package ports
 
 const (
-	OTLPHTTP    = 4318
-	OTLPGRPC    = 4317
-	Metrics     = 8888
-	HealthCheck = 13133
-	Pprof       = 1777
-	IstioEnvoy  = 15090
+	OTLPHTTP    = int32(4318)
+	OTLPGRPC    = int32(4317)
+	Metrics     = int32(8888)
+	HealthCheck = int32(13133)
+	Pprof       = int32(1777)
+	IstioEnvoy  = int32(15090)
 )
